chore move (email)preview from app
