+++ src/docs/components/EmailPreview/PreviewShell.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import * as React from "react";
 import Link from "next/link";
 import { cn } from "@utils/cn";
