+++ src/docs/components/EmailPreview/index.ts
@@ -1,0 +1,4 @@
+export { PreviewNavigation } from "./PreviewNavigation";
+export { PreviewPane } from "./PreviewPane";
+export { PreviewSend } from "./PreviewSend";
+export { PreviewShell } from "./PreviewShell";
